Add files via upload
I have to redesign the "responsiveness" of the general layout so I've took out all the paddings and margins from the css and added bootstrap categories with limits so it change depending on the screen size. Besides that I had to shrink the name on the navbar because it was too big and creating problems.  I don't know how to merge the files so I will do just the pull request.
With these changes it will be working well on phones too.